+++ src/main/java/lotto/io/InputControl.java
@@ -3,9 +3,9 @@ import camp.nextstep.edu.missionutils.Console;
 
 public class InputControl {
 
-    public static long readLong() {
+    public static int readInt() {
         try {
-            return Long.parseLong(readLine());
+            return Integer.parseInt(readLine());
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("[ERROR] 입력한 수가 정수 형식이 아닙니다.");
         }
